make coordinate operation

+++ fdf/srcs/draw.c
@@ -6,7 +6,7 @@
 /*   By: hkubo <hkubo@student.42tokyo.jp>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/08/09 21:38:35 by hkubo             #+#    #+#             */
-/*   Updated: 2021/08/15 22:49:27 by hkubo            ###   ########.fr       */
+/*   Updated: 2021/08/15 22:51:46 by hkubo            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -168,7 +168,7 @@ void	draw_line(t_info *info)
 	int	i;
 	int	j;
 	int	x;
-	int	y;
+	double	y;
 	double	slope;
 
 	i = 0;
@@ -179,12 +179,12 @@ void	draw_line(t_info *info)
 		{
 			slope = (double)(info->map[i][j + 1].y - info->map[i][j].y) / (double)(info->map[i][j + 1].x - info->map[i][j].x);
 			x = info->map[i][j].x;
-			y = info->map[i][j].y;
+			y = (double)info->map[i][j].y;
 			while (x < info->map[i][j + 1].x)
 			{
 				y += slope;
 				x++;
-				my_mlx_pixel_put(&info->img, x, y, 0x00FF0000);
+				my_mlx_pixel_put(&info->img, x, (int)y, 0x00FF0000);
 			}
 			printf("x: %d y: %d slope: %f\n", info->map[i][j + 1].x - info->map[i][j].x, info->map[i][j + 1].y - info->map[i][j].y, slope);
 			j++;
